requirements/development.txt
amqp==2.2.1
asn1crypto==0.22.0
bcrypt==3.1.3
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
cryptography==2.0.3
decorator==4.1.2
Django==1.11.4
django-appconf==1.0.2
django-celery-email==2.0.0
django-extensions==1.8.1
django-filter==1.0.4
djangorestframework==3.6.3
Fabric3==1.13.1.post1
httpie==0.9.9
idna==2.5
ipython==6.1.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.10.2
Jinja2==2.9.6
kombu==4.1.0
MarkupSafe==1.0
paramiko==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3
ptyprocess==0.5.2
pyasn1==0.3.2
pycparser==2.18
Pygments==2.2.0
PyNaCl==1.1.2
pytz==2017.2
requests==2.18.3
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
